requirements.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==2.0.1
django-cors-headers==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-filters==0.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
instamojo-wrapper==1.1
isort==4.2.15
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.0.0
psycopg2==2.7.3.2
pydot==1.2.4
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.4
pyparsing==2.2.0
pytz==2017.3
requests==2.18.4
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
wincertstore==0.2
wrapt==1.10.11
